Move generators to subpackage
